Throttle configuration persistence.

+++ dataplatform/src/main/java/com/latticeengines/dataplatform/dao/impl/ModelDaoImpl.java
@@ -24,7 +24,7 @@ public class ModelDaoImpl extends BaseDaoImpl<Model> implements ModelDao {
     @Override
     public Model deserialize(String id, String content) {
         Model model = new Model();
-        model.setId(Integer.parseInt(id));
+        model.setId(Long.parseLong(id));
         
         if (content != null) {
             String[] jobIds = content.split(",");
